[TASK] Fix Unittests for ProcessService and moved them to correct location
